feat: order item address in api response

+++ payment/models/toss_payment_models.py
@@ -21,7 +21,7 @@ class Order(BaseModel):
     product = models.ForeignKey(
         "product.Products", on_delete=models.CASCADE, related_name="orders"
     )
-    amount = models.DecimalField(max_digits=10, decimal_places=2)
+    amount = models.PositiveIntegerField()
     status = models.CharField(max_length=20, choices=STATUS, default="PENDING")
     payment_key = models.CharField(max_length=255, blank=True, null=True)
     payment_type = models.CharField(max_length=20, blank=True, null=True)
